requirement.txt
langchain
langchain-google-genai
qdrant-client
sentence-transformers
django-redis
spacy
gunicorn
python-dotenv
django-cors-headers
djangorestframework